Sort out the types for extract_patches
This should be the final commit for fixing the types for cython.
I learnt a lot about the interaction between the buffer interfaces
in cython, and thus I also had to add this new cyutils.pxd file.

The only real change here is that the extract_patches method
should now respect the type of the passed in image.