Add new fast piecewise affine code
We need to decide what to do with this - this is a naive
implementation that takes the freely licensed opcode software
and just pops it in Menpo. Perhaps we don't want to something
so heavy handed and we want to roll our own given that we don't
really use alot of the functionality from opcode.

A summary of the issue that this solves:

Give a triangulation and a set of points, find the barycentric
coordinates for the given points. This is particularly important
for dense models that might have 10s of thousands of points
inside the reference frame. To make this fast, this code
basically does a fast ray interesection problem by pretending
that they reference frame is a plane in 3D. Perhaps we just want
to roll our own using an octree or something to make the queries
faster. However, this code is blazingly fast and makes viewing
of dense AAMs possible.