Split PCAModel and PCAInstanceModel
Now you can build models of non-Menpo instances, such as a numpy
array. This is a breaking change - but it should work out easier
for a number of use cases. I have overriden a number of commonly
used methods so that PCAModel shadows an instance backed model,
despite not being one.