Make the landmark manager an attr dict
This is a special kind of mutable mapping that exposes it's
keys as attributes. Uses the correct overriding of
__getattr__ and __setattr_ in order to not override the default
behaviour of the attributes actually on the class (inspired
by Bunch and used under the MIT license:
https://github.com/dsc/bunch/blob/master/LICENSE.txt).

This also makes a stronger enforcement of ascii style string
identifiers only for Python 2 and Unicode identifiers for
Python 3.